refactor: 폴더 구조 변경
- domain 디렉토리 아래에 member, recruit, location과 같이 도메인 별로 폴더 구조를 변경했습니다.
