+++ src/main/java/org/ureca/pinggubackend/domain/apply/entity/Apply.java
@@ -1,18 +1,18 @@
-package org.ureca.pinggubackend.recruit.likes.entity;
+package org.ureca.pinggubackend.domain.apply.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.ureca.pinggubackend.global.entity.BaseEntity;
-import org.ureca.pinggubackend.member.entity.Member;
-import org.ureca.pinggubackend.recruit.entity.Recruit;
+import org.ureca.pinggubackend.domain.member.entity.Member;
+import org.ureca.pinggubackend.domain.recruit.entity.Recruit;
 
 @Entity
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
-public class Likes extends BaseEntity {
+public class Apply extends BaseEntity {
 
     @ManyToOne
     @JoinColumn(name = "member_id")
