+++ src/main/java/org/ureca/pinggubackend/domain/member/enums/MainHand.java
@@ -1,0 +1,6 @@
+package org.ureca.pinggubackend.domain.member.enums;
+
+public enum MainHand {
+    L,
+    R
+}
