+++ src/main/java/org/ureca/pinggubackend/domain/member/enums/Racket.java
@@ -1,4 +1,4 @@
-package org.ureca.pinggubackend.member.enums;
+package org.ureca.pinggubackend.domain.member.enums;
 
 public enum Racket {
     SHACK_HAND,
