--- src/main/java/org/ureca/pinggubackend/member/enums/Gender.java
@@ -1,6 +1,0 @@
-package org.ureca.pinggubackend.member.enums;
-
-public enum Gender {
-    M,
-    F
-}
