--- src/main/java/org/ureca/pinggubackend/member/enums/MainHand.java
@@ -1,6 +1,0 @@
-package org.ureca.pinggubackend.member.enums;
-
-public enum MainHand {
-    L,
-    R
-}
